chore: split up transform/abgeordnetenwatch.py into modules
