+++ src/bundestag/data/transform/abgeordnetenwatch/process.py
@@ -1,5 +1,4 @@
 import json
-import re
 from pathlib import Path
 
 import pandas as pd
@@ -115,34 +114,6 @@ def get_mandates_data(legislature_id: int, path: Path = None) -> pd.DataFrame:
     return df
 
 
-PARTY_PATTERN = re.compile("(.+)\sseit")
-
-
-def extract_party_from_string(s: str) -> str:
-    if not isinstance(s, str):
-        raise ValueError(f"Expected {s=} to be of type string.")
-    elif "seit" in s:
-        return PARTY_PATTERN.search(s).groups()[0]
-    else:
-        return s
-
-
-def get_parties_from_col(
-    row: pd.Series, col="fraction_names", missing: str = "unknown"
-):
-    strings = row[col]
-    if strings is None or not isinstance(strings, list):
-        return [missing]
-    else:
-        return [extract_party_from_string(s) for s in strings]
-
-
-def transform_mandates_data(df: pd.DataFrame) -> pd.DataFrame:
-    df["all_parties"] = df.apply(get_parties_from_col, axis=1)
-    df["party"] = df["all_parties"].apply(lambda x: x[-1])
-    return df
-
-
 def load_vote_json(legislature_id: int, poll_id: int, path: Path = None) -> dict:
     file = data_utils.get_location(
         data_utils.votes_file(legislature_id, poll_id),
@@ -170,7 +141,7 @@ def parse_vote_data(vote: schemas.Vote) -> dict:
     return d
 
 
-def get_votes_df(
+def get_votes_data(
     legislature_id: int,
     poll_id: int,
     path: Path = None,
@@ -211,7 +182,7 @@ def compile_votes_data(legislature_id: int, path: Path = None, validate: bool = 
         total=len(known_id_combos[legislature_id]),
         desc="poll_id",
     ):
-        df = get_votes_df(legislature_id, poll_id, path=path, validate=False)
+        df = get_votes_data(legislature_id, poll_id, path=path, validate=False)
 
         ids = df.loc[df.duplicated(subset=["mandate_id"]), "mandate_id"].unique()
         if len(ids) > 0:
@@ -227,69 +198,3 @@ def compile_votes_data(legislature_id: int, path: Path = None, validate: bool = 
     if validate:
         schemas.VOTES.validate(df_all_votes)
     return df_all_votes
-
-
-def get_politician_names(df: pd.DataFrame, col="mandate"):
-    names = df[col].str.split(" ").str[:-4].str.join(" ")
-    logger.debug(f"Parsing `{col}` to names. Found {names.nunique()} names")
-    return names
-
-
-def transform_votes_data(df: pd.DataFrame) -> pd.DataFrame:
-    df = df.assign(**{"politician name": get_politician_names})
-    return df
-
-
-def run(
-    legislature_id: int,
-    dry: bool = False,
-    raw_path: Path = None,
-    preprocessed_path: Path = None,
-    validate: bool = False,
-) -> pd.DataFrame:
-    logger.info("Start transforming abgeordnetenwatch data")
-
-    if not dry and (raw_path is None or preprocessed_path is None):
-        raise ValueError(
-            f"When {dry=} `raw_path` and or `preprocessed_path` cannot be None."
-        )
-
-    # ensure paths exist
-    if not dry and not raw_path.exists():
-        raise ValueError(f"{raw_path=} doesn't exist, terminating transformation.")
-    if not dry and not preprocessed_path.exists():
-        data_utils.ensure_path_exists(preprocessed_path)
-
-    # polls
-    df = get_polls_data(legislature_id, path=raw_path)
-    if not dry:
-        file = preprocessed_path / f"df_polls_{legislature_id}.parquet"
-        logger.debug(f"writing to {file}")
-        df.to_parquet(path=file)
-
-    # mandates
-    df = get_mandates_data(legislature_id, path=raw_path)
-    df = transform_mandates_data(df)
-
-    if not dry:
-        file = preprocessed_path / f"df_mandates_{legislature_id}.parquet"
-        logger.debug(f"Writing to {file}")
-        df.to_parquet(path=file)
-
-    # votes
-    df_all_votes = compile_votes_data(legislature_id, path=raw_path, validate=validate)
-    df_all_votes = transform_votes_data(df_all_votes)
-
-    if not dry:
-        all_votes_path = (
-            preprocessed_path / f"compiled_votes_legislature_{legislature_id}.csv"
-        )
-        logger.debug(f"Writing to {all_votes_path}")
-
-        df_all_votes.to_csv(all_votes_path, index=False)
-
-        file = preprocessed_path / f"df_all_votes_{legislature_id}.parquet"
-        logger.debug(f"Writing to {file}")
-        df_all_votes.to_parquet(path=file)
-
-    logger.info("Done transforming abgeordnetenwatch data")
